+++ src/common/util/stl_tool.h
@@ -21,7 +21,7 @@ template <typename T>
 std::string Vec2Str(const std::vector<T>& v,
                     const std::function<std::string(const T&)>& f) {
   std::stringstream ss;
-  ss << "vec size = " << v.size() << '\n';
+  ss << "size = " << v.size() << '\n';
   if (!f) return ss.str();
 
   constexpr size_t kMaxLineLen = 32;
@@ -73,7 +73,7 @@ template <typename T>
 std::string Set2Str(const std::set<T>& s,
                     const std::function<std::string(const T&)>& f) {
   std::stringstream ss;
-  ss << "set size = " << s.size() << '\n';
+  ss << "size = " << s.size() << '\n';
   if (!f) return ss.str();
 
   constexpr size_t kMaxLineLen = 32;
@@ -127,7 +127,7 @@ std::string Map2Str(const std::map<KeyType, ValType>& m,
                     const std::function<std::string(const KeyType&)>& fkey,
                     const std::function<std::string(const ValType&)>& fval) {
   std::stringstream ss;
-  ss << "map size = " << m.size() << '\n';
+  ss << "size = " << m.size() << '\n';
   if (!fkey) return ss.str();
 
   constexpr size_t kMaxLineLen = 32;