+++ src/runtime/core/logger/rotate_file_logger_backend.h
@@ -1,9 +1,12 @@
 #pragma once
 
 #include <fstream>
+#include <shared_mutex>
+#include <unordered_map>
 
 #include "aimrt_module_cpp_interface/executor/executor.h"
 #include "core/logger/logger_backend_base.h"
+#include "util/string_util.h"
 
 namespace aimrt::runtime::core::logger {
 
@@ -14,6 +17,7 @@ class RotateFileLoggerBackend : public LoggerBackendBase {
     std::string filename = "aimrt.log";
     uint32_t max_file_size_m = 16;
     uint32_t max_file_num = 100;
+    std::string module_filter = "(.*)";
     std::string log_executor_name = "";
   };
 
@@ -21,15 +25,18 @@ class RotateFileLoggerBackend : public LoggerBackendBase {
   RotateFileLoggerBackend() = default;
   ~RotateFileLoggerBackend() override;
 
-  std::string_view Name() const override { return "rotate_file"; }
+  std::string_view Type() const override { return "rotate_file"; }
 
   void Initialize(YAML::Node options_node) override;
+  void Start() override;
   void Shutdown() override { run_flag_.store(false); }
 
   void RegisterGetExecutorFunc(
       const std::function<aimrt::executor::ExecutorRef(std::string_view)>& get_executor_func) {
     get_executor_func_ = get_executor_func;
   }
+
+  bool AllowDuplicates() const override { return true; }
 
   void Log(const LogDataWrapper& log_data_wrapper,
            const std::shared_ptr<std::string>& format_log_str_ptr) override;
@@ -38,6 +45,7 @@ class RotateFileLoggerBackend : public LoggerBackendBase {
   bool OpenNewFile();
   void CleanLogFile();
   uint32_t GetNextIndex();
+  bool CheckLog(const LogDataWrapper& log_data_wrapper);
 
  private:
   Options options_;
@@ -48,6 +56,11 @@ class RotateFileLoggerBackend : public LoggerBackendBase {
   std::ofstream ofs_;
 
   std::atomic_bool run_flag_ = false;
+
+  std::shared_mutex module_filter_map_mutex_;
+  std::unordered_map<
+      std::string, bool, aimrt::common::util::StringHash, std::equal_to<>>
+      module_filter_map_;
 };
 
 }  // namespace aimrt::runtime::core::logger