+++ src/runtime/core/rpc/rpc_manager.cc
@@ -148,12 +148,17 @@ void RpcManager::Initialize(YAML::Node options_node) {
   // 初始化backend manager
   rpc_backend_manager_.Initialize(rpc_registry_ptr_.get());
 
-  AIMRT_TRACE("Rpc manager init success, backends list: {}",
-              aimrt::common::util::Vec2Str(rpc_backend_name_vec_));
-
   options_node = options_;
 
-  AIMRT_INFO("Rpc init complete.");
+  AIMRT_INFO(R"str(Rpc manager init complete. options:
+----------------------------- aimrt.rpc ----------------------------------------
+{}
+----------------------------- aimrt.rpc ----------------------------------------
+
+rpc backends list: {}
+)str",
+             YAML::Dump(options_node),
+             aimrt::common::util::Vec2Str(rpc_backend_name_vec_));
 }
 
 void RpcManager::Start() {
@@ -162,13 +167,62 @@ void RpcManager::Start() {
       "Function can only be called when state is 'Init'.");
 
   rpc_backend_manager_.Start();
+
+  if (GetLogger().GetLogLevel() <= aimrt::common::util::kLogLevelInfo) {
+    std::vector<std::vector<std::string>> client_info_table =
+        {{"func", "module", "backends"}};
+
+    const auto& client_backend_info = rpc_backend_manager_.GetClientsBackendInfo();
+    const auto& client_index_map = rpc_registry_ptr_->GetClientIndexMap();
+
+    for (const auto& client_index_itr : client_index_map) {
+      auto client_backend_itr = client_backend_info.find(client_index_itr.first);
+      AIMRT_CHECK_ERROR_THROW(client_backend_itr != client_backend_info.end(),
+                              "Invalid rpc registry info.");
+
+      for (const auto& item : client_index_itr.second) {
+        std::vector<std::string> cur_client_info(3);
+        cur_client_info[0] = client_index_itr.first;
+        cur_client_info[1] = item->module_name;
+        cur_client_info[2] = aimrt::common::util::JoinVec(client_backend_itr->second, ",");
+        client_info_table.emplace_back(std::move(cur_client_info));
+      }
+    }
+
+    std::vector<std::vector<std::string>> server_info_table =
+        {{"func", "module", "backends"}};
+
+    const auto& server_backend_info = rpc_backend_manager_.GetServersBackendInfo();
+    const auto& server_index_map = rpc_registry_ptr_->GetServiceIndexMap();
+
+    for (const auto& server_index_itr : server_index_map) {
+      auto server_backend_itr = server_backend_info.find(server_index_itr.first);
+      AIMRT_CHECK_ERROR_THROW(server_backend_itr != server_backend_info.end(),
+                              "Invalid rpc registry info.");
+
+      for (const auto& item : server_index_itr.second) {
+        std::vector<std::string> cur_server_info(3);
+        cur_server_info[0] = server_index_itr.first;
+        cur_server_info[1] = item->module_name;
+        cur_server_info[2] = aimrt::common::util::JoinVec(server_backend_itr->second, ",");
+        server_info_table.emplace_back(std::move(cur_server_info));
+      }
+    }
+
+    AIMRT_INFO(R"str(Rpc manager start complete.
+client info table:{}
+server info table:{}
+)str",
+               aimrt::common::util::DrawTable(client_info_table),
+               aimrt::common::util::DrawTable(server_info_table));
+  }
 }
 
 void RpcManager::Shutdown() {
   if (std::atomic_exchange(&state_, State::Shutdown) == State::Shutdown)
     return;
 
-  AIMRT_INFO("Shutdown rpc.");
+  AIMRT_INFO("Rpc manager Shutdown.");
 
   rpc_handle_proxy_map_.clear();