+++ src/runtime/core/rpc/rpc_registry.h
@@ -49,13 +49,11 @@ class RpcRegistry {
   bool RegisterClientFunc(
       std::unique_ptr<ClientFuncWrapper>&& client_func_wrapper_ptr);
 
-  const auto& GetServiceFuncWrapperMap() const {
-    return service_func_wrapper_map_;
-  }
+  const auto& GetServiceFuncWrapperMap() const { return service_func_wrapper_map_; }
+  const auto& GetClientFuncWrapperMap() const { return client_func_wrapper_map_; }
 
-  const auto& GetClientFuncWrapperMap() const {
-    return client_func_wrapper_map_;
-  }
+  const auto& GetServiceIndexMap() const { return service_index_map_; }
+  const auto& GetClientIndexMap() const { return client_index_map_; }
 
  private:
   std::shared_ptr<aimrt::common::util::LoggerWrapper> logger_ptr_;
@@ -66,10 +64,16 @@ class RpcRegistry {
   using ServicePkgMap = std::unordered_map<std::string_view, ServiceModuleMap>;
   ServicePkgMap service_func_wrapper_map_;
 
+  // 索引表，func_name:wrapper
+  std::unordered_map<std::string_view, std::vector<ServiceFuncWrapper*>> service_index_map_;
+
   // pkg_path:module_name:func_name:wrapper
   using ClientFuncMap = std::unordered_map<std::string_view, std::unique_ptr<ClientFuncWrapper>>;
   using ClientModuleMap = std::unordered_map<std::string_view, ClientFuncMap>;
   using ClientPkgMap = std::unordered_map<std::string_view, ClientModuleMap>;
   ClientPkgMap client_func_wrapper_map_;
+
+  // 索引表，func_name:wrapper
+  std::unordered_map<std::string_view, std::vector<ClientFuncWrapper*>> client_index_map_;
 };
 }  // namespace aimrt::runtime::core::rpc