+++ document/release_notes.md
@@ -14,6 +14,8 @@
 - 提供了每个Module独立的enable开关；
 - 提供了同步RPC接口和异步RPC接口，以及相关example；
 - 原来的协程RPC Proxy接口重命名为CoProxy，之前老的命名将在后几个版本中删除；
+- 优化了框架日志，现在可以在启动日志中查看executor信息、channel/rpc注册信息；
+- consoler/file日志后端现在支持模块过滤功能，并且允许同时注册多个file日志后端，方便将不同模块的日志打印到不同文件中；
 - ros2_plugin：
   - ros2 RPC Backend现在支持非ros2协议；
   - ros2 Channel Backend现在支持非ros2协议；